move meta.json
